Refine native cursor bridge for template use
